list/tree implementation

+++ libc/include/list.h
@@ -24,7 +24,7 @@ void list_destroy( list_t* list );
 void list_free( list_t* list );
 void list_append( list_t* list, node_t* item );
 node_t* list_find( list_t* list, void* value );
-int list_index_of( list_t* list, void* value );
+size_t list_index_of( list_t* list, void* value );
 void list_remove( list_t* list, size_t index );
 void list_delete( list_t* list, node_t* node );
 node_t* list_pop( list_t* list );
